+++ ne_engine/system/SystemThreadPool.hpp
@@ -1,37 +1,12 @@
 #ifndef SYSTEMTHREADPOOL_HPP
 #define SYSTEMTHREADPOOL_HPP
 
+
 #include <functional>
 #include <future>
 #include <queue>
-#include <unordered_set>
 
-
-template<typename NodeType>
-class DAG {
-public:
-	using NodeId = std::size_t;
-
-	NodeId AddNode(NodeType&& n);
-	void AddEdge(NodeId u, NodeId v);
-
-	std::vector<std::size_t> GetInDegrees() const { return std::vector(_in_degrees); }
-	const std::vector<NodeId>& GetNodeNeighbors(NodeId u) const;
-	const NodeType& GetNodeValue(NodeId u) const;
-	const NodeType* GetNodeValuePtr(NodeId u) const;
-
-private:
-	struct EdgeHash {
-		std::size_t operator()(const std::pair<NodeId, NodeId>& e) const {
-			return std::hash<std::size_t>()(e.first) ^ (std::hash<std::size_t>()(e.second) << 1);
-		}
-	};
-
-	std::vector<NodeType> _nodes;
-	std::vector<std::vector<NodeId>> _adjacency_lists;
-	std::unordered_set<std::pair<NodeId, NodeId>, EdgeHash> _edge_set;
-	std::vector<std::size_t> _in_degrees;
-};
+#include "DAG.hpp"
 
 
 class SystemThreadPool {
